Estilização das páginas home.blade, listar-noticias.blade e visualizar.blade, correção de bugs

+++ app/controllers/ComentarioController.php
@@ -24,25 +24,30 @@ class ComentarioController extends BaseController {
 
     public function postEditar() {
         $inputs = Input::all();
+        $comentario = Comentario::where("id", "=", $inputs['comentario_id'])->first();
 
         $validator = Validator::make($inputs, array(
-            "texto" => "required",
+            "texto_comentario" => "required",
             "comentario_id" => "required|exists:comentarios,id"
         ));
 
-        if($validator->passes()) {
-            $comentario = Comentario::where("id", "=", $inputs['comentario_id'])->first();
+        if(Auth::user()->id == $comentario->usuario->id) {
+            if($validator->passes()) {
+                $comentario = Comentario::where("id", "=", $inputs['comentario_id'])->first();
 
-            if (is_null($comentario))
-                return Response::make("Comentário não encontrado.", 404);
+                if (is_null($comentario))
+                    return Response::make("Comentário não encontrado.", 404);
 
-            $comentario->update(array(
-                "texto" => $inputs['texto']
-            ));
+                $comentario->update(array(
+                    "texto" => $inputs['texto_comentario']
+                ));
 
-            return Redirect::back()->with("update_comentario_success", true);
+                return Redirect::back()->with("update_comentario_success", true);
+            } else {
+                return Redirect::back()->withErrors($validator->messages())->withInput();
+            }
         } else {
-            return Redirect::back()->withErrors($validator->messages())->withInput();
+            return Redirect::back()->with("permission_error", true);
         }
     }
 
@@ -50,7 +55,7 @@ class ComentarioController extends BaseController {
         $id = Input::only("comentario_id");
         $comentario = Comentario::where("id", "=", $id)->first();
 
-        if(Auth::user()->id == $comentario->usuario->id) {
+        if(!is_null($comentario) && Auth::user()->id == $comentario->usuario->id) {
             $comentario->delete();
 
             return Redirect::back()->with("delete_comentario_success", true);
